Wai-93: Fix expanded dashboard tiles (#3442)
* refactor component modules

* re style grid

* tweak styles
