+++ packages/web-frontend/src/components/View/View.js
@@ -1,6 +1,6 @@
 /*
  * Tupaia
- * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
+ * Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
  */
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
