--- packages/web-frontend/src/containers/ExpandButton/index.js
@@ -1,10 +1,0 @@
-/**
- * Tupaia Web
- * Copyright (c) 2019 Beyond Essential Systems Pty Ltd.
- * This source code is licensed under the AGPL-3.0 license
- * found in the LICENSE file in the root directory of this source tree.
- */
-
-import ExpandButton from './ExpandButton';
-
-export default ExpandButton;
