+++ packages/web-frontend/src/styles/index.js
@@ -49,7 +49,6 @@ export const MOBILE_HEADER_HEIGHT = 55;
 const MATERIAL_UI_POPOVER_INDEX = 1300; // Taken from material-ui source code
 export const DIALOG_Z_INDEX = MATERIAL_UI_POPOVER_INDEX - 1;
 export const EXPORT_CHART_PADDING = 20;
-export const DASHBOARD_VIEW_MARGIN = 4;
 export const DASHBOARD_META_MARGIN = 15;
 export const TOP_BAR_HEIGHT = 60;
 export const CONTROL_BAR_WIDTH = 340;
@@ -302,7 +301,6 @@ export const VIEW_STYLES = {
         transition: 'min-height 0.3s',
         width: '100%',
         maxWidth: '100%',
-        marginBottom: DASHBOARD_VIEW_MARGIN,
         backgroundColor: '#272832',
         boxSizing: 'border-box',
         flex: '1 0 auto', // fill left over space
@@ -313,13 +311,7 @@ export const VIEW_STYLES = {
         height: '100%',
       }
     : {
-        minHeight: 250,
-        height: 'min-content',
-        maxWidth: 450,
-        minWidth: 300,
         padding: '15px 15px 15px 10px',
-        margin: '0.5%',
-        width: '24%',
         border: '1px solid #1f2038',
         boxShadow: '0px 0px 15px rgba(0,0,0,0.3)',
       },
